// apps/auth-svc/src/proto/auth.ts
export interface UserCreatedEvent {
  userId: string;
  email: string;
  timestamp: number;
  urlToken: string;
}

function createBaseUserCreatedEvent(): UserCreatedEvent {
  return { userId: '', email: '', timestamp: 0, urlToken: '' };
}

export const UserCreatedEvent: MessageFns<UserCreatedEvent> = {
  encode(
    message: UserCreatedEvent,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.userId !== '') {
      writer.uint32(10).string(message.userId);
    }
    if (message.email !== '') {
      writer.uint32(18).string(message.email);
    }
    if (message.timestamp !== 0) {
      writer.uint32(24).int64(message.timestamp);
    }
    if (message.urlToken !== '') {
      writer.uint32(34).string(message.urlToken);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UserCreatedEvent {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUserCreatedEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.email = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.timestamp = longToNumber(reader.int64());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.urlToken = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UserCreatedEvent {
    return {
      userId: isSet(object.userId) ? globalThis.String(object.userId) : '',
      email: isSet(object.email) ? globalThis.String(object.email) : '',
      timestamp: isSet(object.timestamp)
        ? globalThis.Number(object.timestamp)
        : 0,
      urlToken: isSet(object.urlToken)
        ? globalThis.String(object.urlToken)
        : '',
    };
  },

  toJSON(message: UserCreatedEvent): unknown {
    const obj: any = {};
    if (message.userId !== '') {
      obj.userId = message.userId;
    }
    if (message.email !== '') {
      obj.email = message.email;
    }
    if (message.timestamp !== 0) {
      obj.timestamp = Math.round(message.timestamp);
    }
    if (message.urlToken !== '') {
      obj.urlToken = message.urlToken;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UserCreatedEvent>, I>>(
    base?: I,
  ): UserCreatedEvent {
    return UserCreatedEvent.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UserCreatedEvent>, I>>(
    object: I,
  ): UserCreatedEvent {
    const message = createBaseUserCreatedEvent();
    message.userId = object.userId ?? '';
    message.email = object.email ?? '';
    message.timestamp = object.timestamp ?? 0;
    message.urlToken = object.urlToken ?? '';
    return message;
  },
};

type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}